bootctl: split out "install" verb too
