+++ src/boot/bootctl.h
@@ -6,6 +6,20 @@
 #include "json.h"
 #include "pager.h"
 
+typedef enum EntryTokenType {
+        ARG_ENTRY_TOKEN_MACHINE_ID,
+        ARG_ENTRY_TOKEN_OS_IMAGE_ID,
+        ARG_ENTRY_TOKEN_OS_ID,
+        ARG_ENTRY_TOKEN_LITERAL,
+        ARG_ENTRY_TOKEN_AUTO,
+} EntryTokenType;
+
+typedef enum InstallSource {
+        ARG_INSTALL_SOURCE_IMAGE,
+        ARG_INSTALL_SOURCE_HOST,
+        ARG_INSTALL_SOURCE_AUTO,
+} InstallSource;
+
 extern char *arg_esp_path;
 extern char *arg_xbootldr_path;
 extern bool arg_print_esp_path;
@@ -17,11 +31,13 @@ extern bool arg_quiet;
 extern int arg_make_entry_directory; /* tri-state: < 0 for automatic logic */
 extern sd_id128_t arg_machine_id;
 extern char *arg_install_layout;
+extern EntryTokenType arg_entry_token_type;
 extern char *arg_entry_token;
 extern JsonFormatFlags arg_json_format_flags;
 extern bool arg_arch_all;
 extern char *arg_root;
 extern char *arg_image;
+extern InstallSource arg_install_source;
 extern char *arg_efi_boot_option_description;
 
 static inline const char *arg_dollar_boot_path(void) {
